[Brl.LinkedList] Change assignment order in TLink.remove
When removing a link, change neighbour links before setting the link
value to null which is the intuitive order of modifications.
Even though LinkedList is not thread-safe and this change does not
prevent segmentation faults, it may reduce the chance of null access
problems when concurrently modifying and iterating over the list.