Refactors maps to use BRL.Collections (#403)
Refactors the map implementations (IntMap, PtrMap, StringMap, ObjectMap)
to leverage the BRL.Collections library's TreeMap for improved
performance and maintainability. This change eliminates the custom
AVL tree implementation and associated C code.

Adds a `Put` method to the `IMap` interface and `TTreeMap`
implementation to allow updating existing key-value pairs in the map.

Adds a `HasNext` method to `TMapNode` to check if there is a next node.

Adds implementation of Compare for string.